requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langgraph
faiss-cpu
sentence-transformers
chromadb
langchainhub
tavily-python
python-dotenv
wikipedia
google-search-results
langchain_huggingface
langchain-google-genai
langchain-tavily
langchain-groq
grandalf
ipykernel
